requirements.txt
pyzmq
coverage
Sphinx
six
